更新および削除時に以下の3点を修正
  - 対象アイテムが存在しないと404エラーを返す。
  - リクエストボディに含まれるemailと対象アイテム作成者のemailが一致しないと403エラーを返す。
  - 成功時にはレスポンスボディ無しで204を返す。

+++ app/api/item/delete/[id]/route.js
@@ -4,14 +4,31 @@ import { NextResponse } from "next/server"
 import connectDB from "@/app/utils/database";
 import { ItemModel } from "@/app/utils/schemaModels";
 
-export async function DELETE(_, context) {
-    const { id } = context.params
+export async function DELETE(request, context) {
     try {
+        const reqBody = await request.json()
+        const { id } = context.params
+
         await connectDB()
-        await ItemModel.deleteOne({ _id: id })
-        return NextResponse.json({message: "アイテム削除成功"})
+
+        const item = await ItemModel.findById(id)
+        if (!item) {
+            return NextResponse.json(
+                { message: "アイテム削除失敗", detail: "指定されたidのアイテムは存在しません" },
+                { status: 404 }
+            )
+        }
+
+        if (item.email === reqBody.email) {
+            await ItemModel.deleteOne({ _id: id })
+            return new Response(null, { status: 204 })
+        } else {
+            return NextResponse.json(
+                { message: "アイテム削除失敗", detail: "他の人が作成したアイテムです。" },
+                { status: 403 }
+            )
+        }
     } catch (err) {
-        console.error(err)
-        return NextResponse.json({message: "アイテム削除失敗"})
+        return NextResponse.json({ message: "アイテム削除失敗", detail: `${err}` }, { status: 500 })
     }
 }
